move test code into internal package
